Lung cancer in 4 Danish cities 1968-71

This dataset, danish_lung_incidence_df, is a data frame containing counts of incident lung cancer cases and population size in four neighbouring Danish cities by age group from 1968 to 1971.

Usage

data(danish_lung_incidence_df)

Format

A data frame with 24 observations and 4 variables:

city

City of observation (factor with 4 levels)

age

Age group (factor with 6 levels)

pop

Population size (integer)

cases

Number of incident lung cancer cases (integer)

Source

Data taken from the ISwR package version 2.0-10

Details

The dataset name has been kept as 'danish_lung_incidence_df' to avoid confusion with other datasets in the R ecosystem. This naming convention helps distinguish this dataset as part of the PulmoDataSets package. The suffix 'df' indicates that this is a data frame object. The original content has not been modified in any way.